+++ ingress-controller/internal/controllers/gateway/udproute_controller.go
@@ -463,3 +463,8 @@ func (r *UDPRouteReconciler) ensureGatewayReferenceStatusAdded(
 	// the status needed an update and it was updated successfully
 	return true, ctrl.Result{}, nil
 }
+
+// SetLogger sets the logger.
+func (r *UDPRouteReconciler) SetLogger(l logr.Logger) {
+	r.Log = l
+}